+++ harvester/records/__init__.py
@@ -1,4 +1,4 @@
-"""harvester.harvest.records"""
+"""harvester.records"""
 
 # ruff: noqa: I001,F401
 
@@ -9,8 +9,3 @@ from harvester.records.record import (
     XMLSourceRecord,
     JSONSourceRecord,
 )
-from harvester.records.fgdc import FGDC
-from harvester.records.iso19139 import ISO19139
-from harvester.records.gbl1 import GBL1
-from harvester.records.aardvark import Aardvark
-from harvester.records.marc import MARC
