+++ harvester/records/formats/__init__.py
@@ -1,0 +1,9 @@
+"""harvester.records.formats"""
+
+# ruff: noqa: I001,F401
+
+from harvester.records.formats.aardvark import Aardvark
+from harvester.records.formats.fgdc import FGDC
+from harvester.records.formats.iso19139 import ISO19139
+from harvester.records.formats.marc import MARC
+from harvester.records.formats.gbl1 import GBL1
