+++ harvester/records/formats/gbl1.py
@@ -1,8 +1,7 @@
-"""harvester.harvest.records.gbl1"""
+"""harvester.records.formats.gbl1"""
 
 # ruff: noqa: N802
 
-import json
 from typing import Literal
 
 from attrs import define, field
@@ -12,10 +11,7 @@ from harvester.records.record import JSONSourceRecord
 
 @define
 class GBL1(JSONSourceRecord):
-    """GeoBlacklight1.x (GBL1) metadata format SourceRecord class.
-
-    NOTE: This source record class is only used for OGM harvests.
-    """
+    """GeoBlacklight1.x (GBL1) metadata format SourceRecord class."""
 
     metadata_format: Literal["gbl1"] = field(default="gbl1")
 
@@ -48,74 +44,6 @@ class GBL1(JSONSourceRecord):
 
     def _locn_geometry(self) -> str | None:
         return self._dcat_bbox()
-
-    def _dct_references_s_ogm(self) -> dict:
-        """Field method helper: "dct_references_s"
-
-        For most OGM repositories providing GBL1 metadata, pulling this URL from the
-        dct_references_s field directly suffices, and this approach is therefore the
-        default.  However, some repositories require alternate strategies which can be
-        optionally defined in the OGM config YAML using the "external_url_strategy"
-        property.
-
-        Additionally, if the URI "http://schema.org/downloadUrl" is present, and only a
-        single value, use.  If value is array, skip, as we cannot be sure of a single
-        download link to choose from.
-        """
-        # extract required external url
-        url: None | str
-        if external_url_strategy := self.ogm_repo_config.get("external_url_strategy"):
-            url = self._use_external_url_strategy(external_url_strategy)
-        else:
-            refs_dict = json.loads(self.parsed_data["dct_references_s"])
-            url = refs_dict.get("http://schema.org/url")
-        if not url:
-            error_message = "Could not determine external URL from source metadata"
-            raise ValueError(error_message)
-        urls_dict = {"http://schema.org/url": url}
-
-        # extract optional download url
-        download_uri = "http://schema.org/downloadUrl"
-        refs_dict = json.loads(self.parsed_data["dct_references_s"])
-        if download_value := refs_dict.get(download_uri):  # noqa: SIM102
-            if isinstance(download_value, str):
-                urls_dict[download_uri] = [  # type: ignore[assignment]
-                    {
-                        "label": "Data",
-                        "url": download_value,
-                    }
-                ]
-
-        return urls_dict
-
-    def _use_external_url_strategy(self, alternate_strategy: dict) -> str:
-        """Apply alternative strategy for extracting external URL from source record.
-
-        OGM repositories may include an optional "external_url_strategy" property where
-        a sub-property "name" defines the strategy to use.  Currently supported:
-
-            - "base_url_and_slug": a pre-defined base URL is combined with the value from
-            a defined field to construct a URL
-
-            - "field_value": a single field contains a full URL
-        """
-        url: None | str
-        strategy_name = alternate_strategy["name"]
-        if strategy_name == "base_url_and_slug":
-            url = "/".join(
-                [
-                    alternate_strategy["base_url"],
-                    self.parsed_data[alternate_strategy["gbl1_field"]],
-                ]
-            )
-        elif strategy_name == "field_value":
-            url = self.parsed_data.get(alternate_strategy["gbl1_field"])
-            if url and not url.startswith("http"):
-                url = None
-        else:
-            error_message = f"Alternate URL strategy not recognized: {strategy_name}"
-            raise ValueError(error_message)
-        return url
 
     ##########################
     # Optional Field Methods
